[RFR] Fix tests for upstream

+++ cypress/integration/tests/developer/applicationinventory/analysis/upload_binary_analysis.test.ts
@@ -29,10 +29,10 @@ import {
 import { Proxy } from "../../../../models/administrator/proxy/proxy";
 import { Analysis } from "../../../../models/developer/applicationinventory/analysis";
 
-describe("Upload Binary Analysis", { tags: "@tier1" }, () => {
+describe("Upload Binary Analysis", { tags: "@tier4" }, () => {
     before("Login", function () {
         // Prevent hook from running, if the tag is excluded from run
-        if (hasToBeSkipped("@tier1")) return;
+        if (hasToBeSkipped("@tier4")) return;
 
         // Perform login
         login();
@@ -64,6 +64,7 @@ describe("Upload Binary Analysis", { tags: "@tier1" }, () => {
 
     after("Perform test data clean up", function () {
         // Prevent hook from running, if the tag is excluded from run
+        if (hasToBeSkipped("@tier4")) return;
         deleteApplicationTableRows();
         deleteAllBusinessServices();
         writeGpgKey("abcde");
